refactor: Board拡張機能をコマンドとサービスに分離
RedmineCLIプロジェクトと同様の構造に変更し、責任の分離を明確にしました。

変更内容:
- Services/ディレクトリ: ビジネスロジックを分離
  - AuthenticationService: 認証とセッション管理
  - HtmlParsingService: HTML解析ロジック
  - BoardService: ボード関連の操作
- Commands/ディレクトリ: コマンド定義を分離
  - BoardListCommand: ボード一覧コマンド
  - BoardTopicCommand: トピック関連コマンド（動的）
  - InfoCommand: 情報表示コマンド
- Program.cs: DIコンテナの設定とコマンド登録のみに簡素化

効果:
- テストしやすい構造
- 拡張しやすい設計
- RedmineCLIとの一貫性
